langed/requirements.txt
django==5.2.8
uwsgi==2.0.31
djangorestframework==3.15.2
psycopg2_binary==2.9.11
mozilla-django-oidc==4.0.1
Pillow==11.1.0